Updated to make compatible with Java version. Makefile improved for multithread. Tidy popcount() to use SSE4.2 where available.
git-svn-id: http://hdt-it.googlecode.com/svn/trunk@217 ca5865f0-ea67-0c57-7e18-94a3701c88e0

+++ libcds-v1.0.7/src/static/bitsequence/BitSequence.cpp
@@ -130,9 +130,9 @@ namespace cds_static
     }
 
     BitSequence * BitSequence::load(ifstream & fp) {
-        uint r = loadValue<uint>(fp);
+	uchar r = loadValue<uchar>(fp);
         size_t pos = fp.tellg();
-        fp.seekg(pos-sizeof(uint));
+	fp.seekg(pos-sizeof(uchar));
         switch(r) {
             case RRR02_HDR: return BitSequenceRRR::load(fp);
             case BRW32_HDR: return BitSequenceRG::load(fp);
